requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
black==24.1.1
ruff==0.3.5
mypy==1.8.0
types-requests==2.31.0.10
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.14.0
httpx==0.26.0
websockets==15.0.1
python-dotenv==1.0.1
anthropic==0.42.0
